Condensed open and close form down to toggleForm
This fork shows how you can use javascript to check for css styling. Using this method, you only need one function to open or close the hex tab info. (Note: before this, I had a global boolean title isOpen that would toggle true or false whenever the popup opened or closed. The issue with this method is that, in theory, you'd have to create a boolean for each hex (Note within note: just realized this is not entirely true? You could have a global boolean titled hexIsOpen that toggled whenever ANY hex is open and go that route.))